Use the signal2 instead of sigc++
We can use the signal2 instead of sigc++. One of the good thing
is that that the signal2 are thread safe so we can working
from different running loop. Change show how to easily move
from sigc++ to signal2. There can be performance problem on single
process but change the signal to have mutex none in template definition
should avoid this performance penalty

Signed-off-by: Michael Trimarchi <michael@amarulasolutions.com>